[refactor] 폴더 이동한 파일 적용
