+++ src/main/java/klieme/artdiary/calendar/service/CalendarService.java
@@ -12,8 +12,7 @@ import org.springframework.stereotype.Service;
 
 import klieme.artdiary.calendar.enums.CalendarKind;
 import klieme.artdiary.calendar.info.ScheduleInfo;
-import klieme.artdiary.common.ImageTransfer;
-import klieme.artdiary.common.UserIdFilter;
+import klieme.artdiary.common.image.ImageTransfer;
 import klieme.artdiary.exhibition.data_access.entity.ExhEntity;
 import klieme.artdiary.gathering.data_access.entity.GatheringEntity;
 import klieme.artdiary.record_data_access.entity.ExhVisitEntity;
